move version-info to finder
