+++ lib/finder.js
@@ -7,6 +7,7 @@ var debug    = require('debug')('win-detect-browsers')
   , path     = require('path')
   , through2 = require('through2')
   , existent = require('existent')
+  , versionInfo = require('./version-info')
 
 module.exports = Finder
 
@@ -18,6 +19,7 @@ function Finder(name, meta, regStream, opts) {
   this.planned = 0
   this.regStream = regStream
   this.stream = through2.obj()
+  this.readVersion = opts.version
 
   // In lucky mode, end after first result
   if (opts.lucky) this.stream.once('data', this.end.bind(this))
@@ -162,7 +164,13 @@ Finder.prototype.found = function(bin, method, cb, _doTransform) {
 
     // Ignore non-executable's
     if (bin.slice(-4).toLowerCase()==='.exe') {
-      this.stream.write({ name: this.name, path: bin })
+      var b = { name: this.name, path: bin }
+
+      if (this.readVersion) {
+        b = versionInfo(b)
+      }
+
+      if (b) this.stream.write(b)
     }
   }
